requirements.txt
python-telegram-bot==11.1.0
redis==3.2.1
requests~=2.28.0
environs~=9.5.0